properly name admin participant status tag
