+++ spec/components/status_tags/admin_participant_status_tag_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe StatusTags::ParticipantStatusTag, type: :component do
+RSpec.describe StatusTags::AdminParticipantStatusTag, type: :component do
   let!(:participant_profile) { create :ect_participant_profile }
   let(:component) { described_class.new participant_profile: }
